requirements.txt
streamlit
pytube
openai-whisper